+++ src/components/Modals/MultiplePlans/CopyPlanModal.vue
@@ -96,7 +96,7 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import '@/components/Modals/PlanModalDropdown.scss';
+@import '@/components/Modals/MultiplePlans/PlanModalDropdown.scss';
 .selectPlan {
   height: 6rem;
   &-label {
