+++ src/components/Modals/MultiplePlans/EditPlanModal.vue
@@ -84,7 +84,7 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import '@/components/Modals/PlanModalDropdown.scss';
+@import '@/components/Modals/MultiplePlans/PlanModalDropdown.scss';
 
 .editPlan {
   &-delete {
